+++ CHANGELOG.md
@@ -10,6 +10,13 @@ The format is based on [Keep a Changelog], and this project adheres to
 [Keep a Changelog]: https://keepachangelog.com/en/1.0.0/
 [Semantic Versioning]: https://semver.org/spec/v2.0.0.html
 
+## [0.10.2] - 2025-05-29
+
+### Added
+
+- Added `set.NewMarshalingStore()`, which wraps a `set.BinaryStore` with a
+  `marshaler.Marshaler` to store values of arbitrary types.
+
 ## [0.10.1] - 2025-05-29
 
 ### Added